+++ test/mocks/deps-mixed/bueno.js
@@ -1,7 +1,0 @@
-let strings = [
-  'require',
-  'import',
-]
-function hello (friend) {
-  return `Hello, ${friend}!`
-}