+++ test/mocks/deps-mixed/no-bueno.js
@@ -1,3 +1,0 @@
-// No touching!
-import a from 'a'
-let b = require('b')